toolkit: log into amazon ecr public
Without this patch we get too many requests errors from the public
amazon ECR.  If we log in we should have an increased rate limit.

It's not clear if we should use the credential helper or not.  kaniko
supports the ecr-login credential helper here:
https://github.com/awslabs/amazon-ecr-credential-helper?tab=readme-ov-file#configuration

But the https://github.com/aws-actions/amazon-ecr-login also supports
logging into the public repository.

This patch uses the action method.