aws: fix TOOMANYREQUESTS error, log into ecr public
Without this patch kaniko is failing with the following
TOOMANYREQUESTS error:

    INFO[0000] Retrieving image public.ecr.aws/docker/library/golang:1.21-bullseye from registry public.ecr.aws
    error building image: unable to complete operation after 0 attempts, last error: GET https://public.ecr.aws/v2/docker/library/golang/manifests/sha256:301b0f36ff74f5b3b0fcae9a158b6338fd6b6d1ed8231b0fff6460a065cebeb3: TOOMANYREQUESTS: Rate exceeded

This change refactors the jobs.  A new ecr job executes every hour and
sores the docker login credentials into a repository secret.  These
credentials are then read by each job without needing to log into ECR
because the credentials stored in the secret should always be valid and
up to date.